requirement.txt

alpaca-py
pandas
TA-Lib
requests
langchain
langchain-community
